+++ raytracer/ColorTriple.h
@@ -1,26 +1,22 @@
 /*
  * ColorTriple.h
- *
  *  Created on: Oct 10, 2018
  *      Author: clairegoldstein
  */
-
 #ifndef COLORTRIPLE_H_
 #define COLORTRIPLE_H_
-
 #include <string>
-#include <iostream>
 
 using namespace std;
 
 class ColorTriple {
 public:
 	ColorTriple();
-	ColorTriple(const float r,  const float g , const float b);
+	ColorTriple(const double r,  const double g , const double b);
 	int getRed() ;
 	int getGreen() ;
 	int getBlue() ;
-	float CapColor(float color);
+	double CapColor(double color);
 
 	virtual ~ColorTriple();
 	string toString();
